+++ frontend/src/common/mock/getUserInfoSummary/data.ts
@@ -2,7 +2,7 @@ import type { MentoringDetail } from '../../types/MentoringDetail';
 
 export const MENTORING_DETAIL: MentoringDetail = {
   id: 1,
-  mentorName: "김트레이너",
+  mentorName: '김트레이너',
   ratingAverage: ' 3.7',
   ratingCount: 10,
   categories: ['근력 증진', '다이어트', '체형 교정'],
@@ -16,37 +16,37 @@ export const MENTORING_DETAIL: MentoringDetail = {
   chatUrl: '',
   certificates: [
     {
-      certificateId: 1,
+      certificateId: '1',
       title: '스포츠안마 자격증',
       type: 'LICENSE',
       imageUrl: '',
     },
     {
-      certificateId: 2,
+      certificateId: '2',
       title: '한국대학교 졸업증명서',
       type: 'EDUCATION',
       imageUrl: '',
     },
     {
-      certificateId: 3,
+      certificateId: '3',
       title: '헬스 트레이너 자격증',
       type: 'LICENSE',
       imageUrl: '',
     },
     {
-      certificateId: 4,
+      certificateId: '4',
       title: 'PT 자격증',
       type: 'LICENSE',
       imageUrl: '',
     },
     {
-      certificateId: 5,
+      certificateId: '5',
       title: '요가 자격증',
       type: 'LICENSE',
       imageUrl: '',
     },
     {
-      certificateId: 6,
+      certificateId: '6',
       title: '필라테스 자격증',
       type: 'LICENSE',
       imageUrl: '',
