+++ frontend/src/common/mock/getUserInfoSummary/handler.ts
@@ -16,7 +16,7 @@ export const testStateStore = {
 const BASE_URL = process.env.API_BASE_URL;
 const MENTORING_DETAIL_URL = `${BASE_URL}${API_ENDPOINTS.MENTORINGS}/:id`;
 
-const getMentoringDetail = http.get(MENTORING_DETAIL_URL, () => {
+export const getUserInfoSummary = http.get(MENTORING_DETAIL_URL, () => {
   const response = MENTORING_DETAIL;
 
   if (testStateStore.shouldFailRequest) {
@@ -28,5 +28,3 @@ const getMentoringDetail = http.get(MENTORING_DETAIL_URL, () => {
 
   return HttpResponse.json(response, { status: 200 });
 });
-
-export const mentoringDetailHandler = [getMentoringDetail];
