+++ frontend/src/pages/mentoringUpdate/mock/handlers.ts
@@ -1,0 +1,3 @@
+import { getUserInfoSummary } from '../../../common/mock/getUserInfoSummary/handler';
+
+export const mentoringUpdateHandler = [getUserInfoSummary];
